Changed "look" of pages to match instagram better
Come on guys, you didn't really like the sea-monster green, did you? I
changed the header and footer to the same color blue that instagram
uses, the text of the header and footer to white, and the background of
the body to white. Also added Kyle's search bar to index.html.