Added search-results.html and search-results.css
Even though we are using a search bar, it occurs to me that we still
need a results page. I just copied some code from dashboard and added a
"Save Search" button to the top. Currently you cannot navigate to the
search results page from the other pages because I don't have time to
figure out how to link it to the search button right now.